+++ src/Handlers/PlayerSpawnedInWorld.cs
@@ -32,6 +32,7 @@ namespace Amnesia.Handlers
                         HandleStandardRespawnSteps(player);
                         DialogShop.UpdatePrices(player);
                         DialogShop.UpdateMoneyTracker(player.entityId, player.inventory.GetSlots(), player.bag.GetSlots());
+                        PlayerRecord.Load(clientInfo);
                         break;
                     // TODO: case RespawnType.LoadedGame: // local player loading existing game
                     case RespawnType.JoinMultiplayer: // existing player rejoining
@@ -40,6 +41,7 @@ namespace Amnesia.Handlers
                         HandleStandardRespawnSteps(player);
                         DialogShop.UpdatePrices(player);
                         DialogShop.UpdateMoneyTracker(player.entityId, player.inventory.GetSlots(), player.bag.GetSlots());
+                        PlayerRecord.Load(clientInfo);
                         break;
                     case RespawnType.Died: // existing player returned from death
                         _ = PlayerHelper.AddPositiveOutlookTime(player, Config.PositiveOutlookTimeOnMemoryLoss);